+++ sqlstore/src/it/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernateLibraryDaoSearchIT.java
@@ -1,0 +1,31 @@
+package uk.ac.bbsrc.tgac.miso.persistence.impl;
+
+import java.util.Arrays;
+import java.util.EnumSet;
+import java.util.List;
+
+import uk.ac.bbsrc.tgac.miso.HibernatePaginatedDataSourceIT;
+
+public class HibernateLibraryDaoSearchIT extends HibernatePaginatedDataSourceIT {
+
+  private static final EnumSet<SearchType> VALID_SEARCH_TYPES = EnumSet.of(SearchType.QUERY, SearchType.ID, SearchType.IDS,
+      SearchType.FREEZER, SearchType.DISTRIBUTED, SearchType.DISTRIBUTION_RECIPIENT, SearchType.TISSUE_ORIGIN, SearchType.TISSUE_TYPE,
+      SearchType.PROJECT, SearchType.BOX, SearchType.CREATED, SearchType.UPDATED, SearchType.RECEIVED, SearchType.WORKSET,
+      SearchType.CREATOR, SearchType.MODIFIER, SearchType.BULK_LOOKUP, SearchType.ENTERED, SearchType.GROUP_ID, SearchType.INDEX,
+      SearchType.KIT_NAME, SearchType.PLATFORM_TYPE, SearchType.WORKSTATION);
+
+  private static final List<String> SORT_FIELDS = Arrays.asList("effectiveTissueOriginLabel", "effectiveTissueTypeLabel", "parentSampleId",
+      "parentSampleAlias");
+
+  public HibernateLibraryDaoSearchIT() {
+    super(VALID_SEARCH_TYPES, SORT_FIELDS);
+  }
+
+  @Override
+  protected HibernatePaginatedDataSource<?> constructTestSubject() {
+    HibernateLibraryDao sut = new HibernateLibraryDao();
+    sut.setSessionFactory(getSessionFactory());
+    return sut;
+  }
+
+}